moves static metadata util methods to only places using them
While investigating #3858, noticed these utility metadata utility
methods that were not widely used.  Moved the static method where
they are actually used to clean up the code.
